Made custom components for sockets
Made custom static mesh components for the sockets and changed their colors.

I still havent finished the sockets with all the building blocks because I cant even get the foundations to work properly. currently trying to figure out how to prevent socket overlapping using a checker imbedded in the socket that deletes a socket when overlapping with another